move Spinner styles to CSS module + move to a new component file
