+++ src/components/index.ts
@@ -2,8 +2,9 @@ import Breadcrumb from './Breadcrumb.js'
 import Cell from './Cell.js'
 import File from './File.js'
 import Folder from './Folder.js'
-import Layout, { ErrorBar, Spinner } from './Layout.js'
+import Layout, { ErrorBar } from './Layout.js'
 import Markdown from './Markdown.js'
 import Page from './Page.js'
+import Spinner from './Spinner.js'
 export * from './viewers/index.js'
 export { Breadcrumb, Cell, ErrorBar, File, Folder, Layout, Markdown, Page, Spinner }
