+++ src/components/viewers/ImageView.tsx
@@ -3,7 +3,7 @@ import { useConfig } from '../../hooks/useConfig.js'
 import { FileSource } from '../../lib/sources/types.js'
 import { cn, contentTypes, parseFileSize } from '../../lib/utils.js'
 import styles from '../../styles/viewers/ImageView.module.css'
-import { Spinner } from '../Layout.js'
+import Spinner from '../Spinner.js'
 import ContentWrapper from './ContentWrapper.js'
 
 interface ViewerProps {
