+++ src/components/viewers/ParquetView.tsx
@@ -1,4 +1,5 @@
 import HighTable, { DataFrame, rowCache } from 'hightable'
+import 'hightable/src/HighTable.css'
 import { asyncBufferFromUrl, parquetMetadataAsync } from 'hyparquet'
 import React, { useCallback, useEffect, useState } from 'react'
 import { useConfig } from '../../hooks/useConfig.js'
@@ -7,7 +8,7 @@ import { FileSource } from '../../lib/sources/types.js'
 import { parquetDataFrame } from '../../lib/tableProvider.js'
 import { cn } from '../../lib/utils.js'
 import styles from '../../styles/ParquetView.module.css'
-import { Spinner } from '../Layout.js'
+import Spinner from '../Spinner.js'
 import CellPanel from './CellPanel.js'
 import ContentWrapper, { ContentSize } from './ContentWrapper.js'
 import SlidePanel from './SlidePanel.js'
